improved precision

+++ Options.cpp
@@ -22,7 +22,7 @@ void Options::update()
 // Wyświetlamy opcje
 void Options::render()
 {
-    DrawText("Press F to enable dark theme", ((float)800 / 2) - MeasureText("Press F to enable dark theme", 30) / 2, 800 / 2, 30, BLACK);
-    DrawText("Press M to mute game", ((float)800 / 2) - MeasureText("Press M to mute game", 30) / 2, 800 / 2 + 40, 30, BLACK);
-    DrawText("Press BACKSPACE to back to menu", ((float)800 / 2) - MeasureText("Press BACKSPACE to back to menu", 20) / 2, (800 / 2) + 90, 20, BLACK);
+    DrawText("Press F to enable dark theme", ((float)800 / 2) - (float)MeasureText("Press F to enable dark theme", 30) / 2, 800 / 2, 30, BLACK);
+    DrawText("Press M to mute game", ((float)800 / 2) - (float)MeasureText("Press M to mute game", 30) / 2, 800 / 2 + 40, 30, BLACK);
+    DrawText("Press BACKSPACE to back to menu", ((float)800 / 2) - (float)MeasureText("Press BACKSPACE to back to menu", 20) / 2, (800 / 2) + 90, 20, BLACK);
 }
